moved `Providers` to fix linting
